+++ processor/src/test/resources/input/com/example/inject/DaggerTrueComponent.java
@@ -8,7 +8,7 @@ import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
-@ReactComponent( dagger = ReactComponent.Feature.TRUE )
+@ReactComponent( dagger = ReactComponent.Feature.ENABLE )
 class DaggerTrueComponent
   extends Component<BaseProps, BaseState, BaseContext>
 {