+++ processor/src/test/resources/input/com/example/inject/InjectFalseComponent.java
@@ -8,7 +8,7 @@ import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
-@ReactComponent( inject = ReactComponent.Feature.FALSE )
+@ReactComponent( inject = ReactComponent.Feature.DISABLE )
 class InjectFalseComponent
   extends Component<BaseProps, BaseState, BaseContext>
 {