+++ processor/src/test/resources/input/com/example/inject/InjectTrueComponent.java
@@ -8,7 +8,7 @@ import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
-@ReactComponent( inject = ReactComponent.Feature.TRUE )
+@ReactComponent( inject = ReactComponent.Feature.ENABLE )
 class InjectTrueComponent
   extends Component<BaseProps, BaseState, BaseContext>
 {